Test of Server Connection.
Created code to test if server is connected. Returns the least damaging type of table just in case this ever gets into production. But it better not ever get called by a user!

Also this demonstrates that the problem noted in previous commit is NOT a result of bad server-api connection.